Refactor DatabaseLoader by splitting it into DatabaseOrchestrator and AtomPersister.
