+++ src/db_adapter/AtomPersister.cc
@@ -1,9 +1,4 @@
-#include "DatabaseLoader.h"
-
-#include <filesystem>
-
-#include "ContextLoader.h"
-#include "Utils.h"
+#include "AtomPersister.h"
 
 #define LOG_LEVEL INFO_LEVEL
 #include "Logger.h"
@@ -14,146 +9,17 @@ using namespace commons;
 using namespace atoms;
 using namespace db_adapter;
 
-namespace fs = std::filesystem;
-
 // ==============================
 //  Construction / destruction
 // ==============================
 
-DatabaseMappingOrchestrator::DatabaseMappingOrchestrator(const JsonConfig& config,
-                                                         shared_ptr<BoundedSharedQueue> output_queue) {
-    this->database_setup(config, output_queue);
-    this->tasks = this->strategy->create_tasks(config);
-}
-
-DatabaseMappingOrchestrator::~DatabaseMappingOrchestrator() { this->db_conn->stop(); }
-
-// ==============================
-//  Public
-// ==============================
-
-bool DatabaseMappingOrchestrator::thread_one_step() {
-    LOG_DEBUG("DatabaseMappingOrchestrator thread_one_step called. Current task index: "
-              << this->current_task);
-    if (this->current_task >= this->tasks.size()) {
-        this->db_conn->stop();
-        return false;
-    }
-
-    if (!this->initialized) {
-        this->db_conn->setup();
-        this->db_conn->start();
-        this->initialized = true;
-    }
-
-    auto& task = this->tasks[this->current_task];
-
-    LOG_DEBUG("Processing task " << task.task_name);
-
-    this->strategy->execute_task(task);
-
-    this->current_task++;
-    this->finished = (this->current_task >= this->tasks.size());
-    return !this->finished;
-}
-
-bool DatabaseMappingOrchestrator::is_finished() const { return this->finished; }
-
-// ==============================
-//  Private
-// ==============================
-
-void DatabaseMappingOrchestrator::database_setup(const JsonConfig& config,
-                                                 shared_ptr<BoundedSharedQueue> output_queue) {
-    string adapterdb_type = config.at_path("adapterdb.type").get_or<string>("");
-
-    // TODO: In the future we can implement a factory pattern here to support more database types
-    if (adapterdb_type == "postgres") {
-        string host = config.at_path("adapterdb.database_credentials.host").get<string>();
-        uint port = config.at_path("adapterdb.database_credentials.port").get<uint>();
-        string username = config.at_path("adapterdb.database_credentials.username").get<string>();
-        string password = config.at_path("adapterdb.database_credentials.password").get<string>();
-        string database = config.at_path("adapterdb.database_credentials.database").get<string>();
-        auto conn = make_unique<PostgresDatabaseConnection>(
-            "psql-conn", host, port, database, username, password);
-        auto postgres_wrapper = make_unique<PostgresWrapper>(*conn, output_queue);
-        auto postgres_strategy = make_unique<PostgresMappingStrategy>(move(postgres_wrapper));
-        this->db_conn = move(conn);
-        this->strategy = move(postgres_strategy);
-    } else {
-        RAISE_ERROR("Unsupported adapter type: " + adapterdb_type);
-    }
-}
-
-PostgresMappingStrategy::PostgresMappingStrategy(unique_ptr<PostgresWrapper> wrapper)
-    : wrapper(move(wrapper)) {}
-
-vector<MappingTask> PostgresMappingStrategy::create_tasks(const JsonConfig& config) {
-    vector<string> file_paths =
-        config.at_path("adapterdb.context_mapping_paths").get_or<vector<string>>({});
-
-    vector<MappingTask> tasks;
-
-    if (file_paths.empty()) {
-        LOG_INFO(
-            "No context mapping files specified in config at adapterdb.context_mapping_paths. The "
-            "entire database will be mapped.");
-        tasks.push_back(MappingTask{"map_all_database", nullopt});
-    } else {
-        for (const auto& path : file_paths) {
-            fs::path p(path);
-            string ext = p.extension().string();
-
-            if (ext != ".sql") {
-                RAISE_ERROR("Unsupported mapping file type: " + ext + " for file: " + path);
-            }
-
-            auto queries_sql = ContextLoader::load_query_file(path);
-
-            if (!queries_sql.empty()) {
-                for (size_t i = 0; i < queries_sql.size(); i++) {
-                    LOG_DEBUG("Query " << (i + 1) << ": " << queries_sql[i]);
-                    tasks.push_back(MappingTask{"custom_query_" + to_string(i), queries_sql[i]});
-                }
-            }
-
-            LOG_DEBUG(to_string(queries_sql.size()) + " queries were loaded from the query file.");
-        }
-    }
-    return tasks;
-}
-
-void PostgresMappingStrategy::execute_task(const MappingTask& task) {
-    if (task.context == nullopt) {
-        auto tables = this->wrapper->list_tables();
-
-        if (tables.empty()) {
-            RAISE_ERROR("No tables found in the database.");
-        }
-
-        for (const auto& table : tables) {
-            this->wrapper->map_table(table, vector<string>{}, vector<string>{}, false);
-        }
-    } else {
-        this->wrapper->map_sql_query(task.task_name, task.context.value());
-    }
-}
-
-/**
- * MultiThreadAtomPersister implementation using ThreadPool
- */
-
-// ==============================
-//  Construction / destruction
-// ==============================
-
-MultiThreadAtomPersister::MultiThreadAtomPersister(shared_ptr<BoundedSharedQueue> input_queue,
-                                                   ThreadPool& pool,
-                                                   shared_ptr<AtomDB> atomdb,
-                                                   size_t batch_size,
-                                                   bool save_metta_expression,
-                                                   string metta_output_dir,
-                                                   size_t max_pending_batches)
+AtomPersister::AtomPersister(shared_ptr<BoundedSharedQueue> input_queue,
+                             ThreadPool& pool,
+                             shared_ptr<AtomDB> atomdb,
+                             size_t batch_size,
+                             bool save_metta_expression,
+                             string metta_output_dir,
+                             size_t max_pending_batches)
     : input_queue(input_queue),
       pool(pool),
       atomdb(atomdb),
@@ -166,17 +32,17 @@ MultiThreadAtomPersister::MultiThreadAtomPersister(shared_ptr<BoundedSharedQueue
         this->metta_writer = make_shared<MettaFileWriter>(metta_output_dir);
     }
 
-    LOG_DEBUG("MultiThreadAtomPersister initialized | batch_size: "
-              << batch_size << " | max_pending_batches: " << max_pending_batches
-              << " | pool: " << this->pool.to_string());
+    LOG_DEBUG("AtomPersister initialized | batch_size: " << batch_size << " | max_pending_batches: "
+                                                         << max_pending_batches
+                                                         << " | pool: " << this->pool.to_string());
 }
 
-MultiThreadAtomPersister::~MultiThreadAtomPersister() {
+AtomPersister::~AtomPersister() {
     if (is_save_metta()) {
         this->metta_writer->close();
     }
 
-    LOG_DEBUG("MultiThreadAtomPersister destroyed | total_atoms: "
+    LOG_DEBUG("AtomPersister destroyed | total_atoms: "
               << this->total_count.load() << " | batches_dispatched: " << this->batches_dispatched.load()
               << " | batches_completed: " << this->batches_completed.load()
               << " | batches_failed: " << this->batches_failed.load());
@@ -186,7 +52,7 @@ MultiThreadAtomPersister::~MultiThreadAtomPersister() {
 //  Public
 // ==============================
 
-void MultiThreadAtomPersister::dispatch() {
+void AtomPersister::dispatch() {
     if (static_cast<size_t>(this->pool.size()) >= this->max_pending_batches) {
         this->flush_batch();
         return;
@@ -213,7 +79,7 @@ void MultiThreadAtomPersister::dispatch() {
     }
 }
 
-void MultiThreadAtomPersister::set_producer_finished() {
+void AtomPersister::set_producer_finished() {
     this->producer_finished.store(true);
     LOG_DEBUG("Producer finished signal received"
               << " | accumulator_size: " << this->accumulator.size() << " | queue_remaining: "
@@ -222,15 +88,15 @@ void MultiThreadAtomPersister::set_producer_finished() {
               << " | batches_completed: " << this->batches_completed.load());
 }
 
-bool MultiThreadAtomPersister::is_producer_finished() const { return this->producer_finished.load(); }
+bool AtomPersister::is_producer_finished() const { return this->producer_finished.load(); }
 
-int MultiThreadAtomPersister::get_total_count() const { return this->total_count.load(); }
+int AtomPersister::get_total_count() const { return this->total_count.load(); }
 
 // ==============================
 //  Private
 // ==============================
 
-void MultiThreadAtomPersister::drain_into_accumulator() {
+void AtomPersister::drain_into_accumulator() {
     size_t limit = (this->accumulator.size() < this->batch_size)
                        ? (this->batch_size - this->accumulator.size())
                        : 0;
@@ -259,7 +125,7 @@ void MultiThreadAtomPersister::drain_into_accumulator() {
     }
 }
 
-void MultiThreadAtomPersister::flush_batch() {
+void AtomPersister::flush_batch() {
     while (this->accumulator.size() >= this->batch_size) {
         if (static_cast<size_t>(this->pool.size()) >= this->max_pending_batches) {
             LOG_DEBUG("flush_batch() | pool_pending_tasks: "
@@ -286,9 +152,7 @@ void MultiThreadAtomPersister::flush_batch() {
     }
 }
 
-void MultiThreadAtomPersister::send_batch(vector<Atom*> atoms,
-                                          int batch_id,
-                                          shared_ptr<MettaFileWriter> writer) {
+void AtomPersister::send_batch(vector<Atom*> atoms, int batch_id, shared_ptr<MettaFileWriter> writer) {
     StopWatch timer_success;
     StopWatch timer_failure;
     timer_success.start();
@@ -317,9 +181,10 @@ void MultiThreadAtomPersister::send_batch(vector<Atom*> atoms,
         }
 
         timer_success.stop();
-
+#if LOG_LEVEL >= DEBUG_LEVEL
         int new_total =
             this->total_count.fetch_add(static_cast<int>(atoms.size())) + static_cast<int>(atoms.size());
+#endif
         this->batches_completed.fetch_add(1);
 
         LOG_DEBUG("Batch #" << batch_id << " completed | size: " << atoms.size()
